+++ CHANGELOG.md
@@ -19,11 +19,31 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Deployment Migration Notes
 
+## [v1.1.1] - 2023-10-11
+
+### Overview
+
+Bugfix release for the asset properties issue.
+
+### Detailed Changes
+
+#### Patch
+
+- Fixed a bug causing some string asset properties getting quotes around them.
+
+### Deployment Migration Notes
+
+#### Compatible Versions
+
+- Broker Backend Docker Image: `ghcr.io/sovity/broker-server-ce:1.1.1`
+- Broker UI Docker Image: `ghcr.io/sovity/edc-ui:0.0.1-milestone-8-sovity13`
+- Sovity EDC CE: [`4.2.0`](https://github.com/sovity/edc-extensions/tree/v4.2.0/connector)
+
 ## [v1.1.0] - 2023-09-29
 
 ### Overview
 
-Bugfix release for the asset proprties issue. Also contains the connector delete endpoint.
+Bugfix release for the asset properties issue. Also contains the connector delete endpoint.
 
 ### Detailed Changes